requirements.txt
asgiref==3.7.2
cloudinary==1.36.0
Django==3.2.22
django-cloudinary-storage==0.3.0
django-filter==23.3
django-multiupload==0.6.1
djangorestframework==3.14.0
Pillow==8.2.0
pytz==2023.3.post1
sqlparse==0.4.4
